experiments/requirements.txt
Cython
jupyter
numpy
pandas
pyscf

# PyTorch (choose a compatible version from the list below)
# -f https://download.pytorch.org/whl/torch_stable.html
# torch==1.8.1+cpu  
# torch==1.9.0+cpu 
# torch
torchvision

pennylane

qulacs-gpu
PennyLane-qulacs

PennyLane-qiskit
qiskit==0.45.3
qiskit-aer
qiskit-aqua==0.8.2
qiskit-ibmq-provider==0.11.1
qiskit-ignis==0.5.2
qiskit-terra
scikit-learn
scipy
seaborn==0.10.1
sympy==1.7.1
pyeda
tensorboard

pathos
pyyaml
tqdm
